👷 truncate long ALIAS_DOMAINS values
